requirements.txt
pandas
mysql-connector-python
python-dotenv
sqlalchemy